requirements.txt
flask>=3.0.0
google-generativeai>=0.7.0
pypdf>=4.2.0
pdf2image>=1.17.0
pytesseract>=0.3.10
python-dotenv>=1.0.1
